+++ frontend/src/shared/library/hooks/index.ts
@@ -2,3 +2,4 @@ export { useMediaQuery } from './use-media-query'
 export { useQueryParams } from './use-query-params'
 export { useStrictContext, createStrictContext } from './use-strict-context'
 export { useDebounce } from './use-debounce.ts'
+export { useRefreshEmailVerified } from './use-refresh-email-verified.ts'
